requirements.txt
fastapi
uvicorn[standard]
boto3
openai
google-cloud-texttospeech